requirements.txt
python-telegram-bot==20.3
groq==0.4.1
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
